fix(01-26): review 7d96a98 companion fixes per operator disposition (WINDOWS 17-19)
- WR-01 (image_tx_manager.cpp): enqueueCapture full-buffer ps_malloc failure falls through to the thumbnail branch instead of early-returning — full marked unavailable (nullptr/0/0) with the renamed 'full dropped, thumbnail still pushes' log; oversize-skip, thumb-alloc-failure, and both-empty guard branches untouched
- WR-02 (main_balloon.cpp): honest camera health check — active camera warns only when esp_camera_sensor_get() returns nullptr (message names the absent handle); inactive camera gets the neighboring 'health check skipped' SYS_WARNING shape; healthy active camera warns nothing
- WR-03 (main_basestation.cpp): ackedAtLastPoll widened to uint32, both truncating static_cast<uint16_t> calls deleted from the LED-truth poll site; counter, poll cadence, staleness thresholds untouched
- disposition: option-a fix-all, auto-selected at the Task 2 gate=blocking checkpoint under workflow.auto_advance (01-23 convention) — recorded for end-of-phase operator confirmation
- WINDOWS 17-19 flipped fixed in both copies with evidence classes: 17/19 closed on code + builds 2/2 + harness exit 0 with unstageable triggers recorded by name (on-demand PSRAM exhaustion / 65536-ACK wrap); 18 names its stageable 01-27 healthy-boot discriminator; counts 3 open / 16 fixed / 19 total

+++ src/main_basestation.cpp
@@ -78,7 +78,7 @@ struct BaseStationState {
     char lastStatus[64];
     // LED truth (IN-03): link state derived from real ACK activity
     uint32_t lastAckTime;
-    uint16_t ackedAtLastPoll;
+    uint32_t ackedAtLastPoll;
     bool lastOutcomeBad;
     uint32_t lastTerminalFailTime;
     // Auto-capture chip: latched from the newest ACKed auto-capture command
@@ -2202,12 +2202,14 @@ void processLoRa() {
         }
     }
 
-    // LED truth (IN-03): record when the latest ACK arrived
+    // LED truth (IN-03): record when the latest ACK arrived — uint32
+    // end-to-end (WR-03, review 7d96a98): no truncating cast, so a
+    // 65536-ACK wrap between polls can no longer skip this refresh
     uint32_t acked = CmdSender().getCommandsAcked();
-    if (static_cast<uint16_t>(acked) > appState.ackedAtLastPoll) {
+    if (acked > appState.ackedAtLastPoll) {
         appState.lastAckTime = millis();
     }
-    appState.ackedAtLastPoll = static_cast<uint16_t>(acked);
+    appState.ackedAtLastPoll = acked;
 }
 
 // ===========================
